cgal depends on CImg
