+++ applications/plugins/CImgPlugin/CMakeLists.txt
@@ -4,7 +4,8 @@ project(CImgPlugin VERSION 0.1)
 set(HEADER_FILES
     src/CImgPlugin/ImageCImg.h
     src/CImgPlugin/CImgPlugin.h
-    src/CImgPlugin/SOFACImg.h
+	src/CImgPlugin/SOFACImg.h
+	src/CImgPlugin/CImgData.h
     )
 
 set(SOURCE_FILES
@@ -86,7 +87,7 @@ add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
 
 target_compile_options(${PROJECT_NAME} PUBLIC ${CIMG_CFLAGS})
 target_compile_options(${PROJECT_NAME} PRIVATE "-DSOFA_BUILD_CIMGPLUGIN")
-target_link_libraries(${PROJECT_NAME} SofaCore ${EXTERNAL_LIBS})
+target_link_libraries(${PROJECT_NAME} SofaCore SofaSimulationCommon ${EXTERNAL_LIBS})
 target_include_directories(${PROJECT_NAME} PUBLIC "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>")
 target_include_directories(${PROJECT_NAME} PUBLIC "$<BUILD_INTERFACE:${CImg_INCLUDE_DIRS}>")
 target_include_directories(${PROJECT_NAME} PUBLIC "$<INSTALL_INTERFACE:include>")
